Update font type row
